requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.1
async-generator==1.10
attrs==25.4.0
blinker==1.6.2
boto3==1.40.56
botocore==1.40.76
branca==0.8.2
brotli==1.2.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
folium==0.20.0
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kaitaistruct==0.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mypy_extensions==1.1.0
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==2.4.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.1
playwright==1.58.0
prompt_toolkit==3.0.52
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyogrio==0.12.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyproj==3.7.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
s3transfer==0.14.0
ruff==0.11.12
selenium==4.40.0
selenium-wire==5.1.0
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.54.0
streamlit-folium==0.26.1
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wheel==0.46.3
wsproto==1.3.2
xyzservices==2025.11.0
zipp==3.23.0
zstandard==0.25.0
